Updated and Created Files
Created a new html page called "cart"
contains the checkout page of the website.

Updated index.html to contain a shopping cart option, also added a row at the top containing the title of the project

Updated index.js to host the backend functionality of the checkout page specifically the quantity up and down buttons.

Lastly, updated the product.html to contain the shopping cart option, also added the functionality of quantity amount.